layout/generic/test/test_bug527306.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=527306
-->
<head>
  <title>Test for Bug 527306</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

</head>
<body onload="runTests()">
<p><a target="_blank" href="https://bugzilla.mozilla.org/show_bug?id=527306">Mozilla Bug 527306</a>

<p><input type="text" id="t1" value="FAIL"></p>
<p><input type="text" id="t2" value="FAIL"></p>
<p><input type="text" id="t3" value="FAIL"></p>
<p><textarea id="t4">FAIL</textarea></p>

<pre id="test">
<script>

function testElement(t) {
  t.style.display = "none";
  t.value = "PASS";
  is(t.value, "PASS", "Value should be set correctly");
}

function runTests() {
  var t = document.getElementById("t1");
  testElement(t);
  t = document.getElementById("t2");
  t.focus();
  testElement(t);
  t = document.getElementById("t3");
  t.focus();
  t.blur();
  testElement(t);
  t = document.getElementById("t4");
  testElement(t);
  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();
</script>
</pre>

</body>
</html>
